% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateFeatureSelection-package.R
\docType{package}
\name{EvaluateFeatureSelection-package}
\alias{EvaluateFeatureSelection}
\alias{EvaluateFeatureSelection-package}
\title{EvaluateFeatureSelection: Evaluate Feature Selection Methods Using Supervised Learning}
\description{
This package generates plots to visualize and assess the performance of feature selection methods using supervised learning. It also provides functions to plot scree plots to visualize good cutting points for the number of features to be selected.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ahmad-Alsaleh/EvaluateFeatureSelection}
  \item \url{https://ahmad-alsaleh.github.io/EvaluateFeatureSelection/}
  \item Report bugs at \url{https://github.com/Ahmad-Alsaleh/EvaluateFeatureSelection/issues}
}

}
\author{
\strong{Maintainer}: Ahmad Alsaleh \email{ahmed.asaleh2@gmail.com}

}
\keyword{internal}
