% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auc_plot.R
\name{get_auc_plot}
\alias{get_auc_plot}
\title{Generate AUC plots to assess the performance of feature selection methods}
\usage{
get_auc_plot(dataset, imp.features.names)
}
\arguments{
\item{dataset}{(data.frame) the dataset to be used. Last column must be the response variable.}

\item{imp.features.names}{(data.frame) each column contains the names of the features names sorted from most to least important}
}
\value{
(gtable) a plot with the AUCs of the different models
}
\description{
Generate AUC plots to assess the performance of feature selection methods
}
