% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateFeatureSelection-package.R
\docType{package}
\name{EvaluateFeatureSelection-package}
\alias{EvaluateFeatureSelection}
\alias{EvaluateFeatureSelection-package}
\title{EvaluateFeatureSelection: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ahmad-Alsaleh/EvaluateFeatureSelection}
  \item \url{https://ahmad-alsaleh.github.io/EvaluateFeatureSelection/}
  \item Report bugs at \url{https://github.com/Ahmad-Alsaleh/EvaluateFeatureSelection/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
